requirements.txt
tensorflow-gpu==2.5
scikit-learn
numpy
google
pandas
gcsfs